server: Added SSL negotiation
A client may ask if the server supports SSL through a special
StartupMessage. This is now supported and was the cause of psql hanging
on connect.

Added a "-v" option to the server (disabled by default) to hide all of
the sent and received messages (which may be helpful for debugging).

Added a "-p" option to set the server port.

Fixes #29