Prepared statements
A prepared statement is compiled and validated, but not executed. It
can then be executed with a set of host parameters (in the form of
":name") to be substituted into the statement. Each invocation
requires all host parameters to be passed in.

The Connection now offers a prepare() function, which can be
subsequently executed using query() with some (or none) host
parameters. The existing Connection.query() now simply acts as a proxy
for a prepared statement without any host parameters.

Added SQLSTATE 42P02 "parameter does not exist".