Added a query cache (#35)
The query cache provides tooling to cache previously parsed prepared
statements. This is becuase parsing a statement is extremely expensive
with the current Earley implementation and many queries (excluding
values) are used more than once.

The query cache is made more useful by the fact it can turn any
existing query into a prepared statement so that cache works in all
cases.

A new query cache is provided with a connection. However, you can also
share a query cache between multiple connections. This is done in tests
to both speed up the tests and to battle test the query cache with a
lot of queries over many connections.

Also fixed a bug where host parameters could not be used in FETCH and
OFFSET clauses.