Adding "bench" tool
A new command "vsql bench" can now be used to benchmark the performance
of vsql. This is useful as a local tool and provides insides to changes
in performance between vsql versions. The TPC-B-like commands are
inspired/copied from pgbench.

See https://github.com/elliotchance/vsql/blob/main/docs/benchmark.rst

This also adds elapsed parsing and execution times to a Result to help
with more precise benchmarking in the future.