Changing storage format
This is a radical change to the storage format that converts each object
(row or table) to being encoded and decoded by the entity itself rather
than the storage itself. This change is required in preparation for the
next major release that will use a B-tree on disk rather than the list
of objects.