Rewrite file format to be a B-tree
This is an extensive refactoring and replacement of the file format to
store all objects (rows and table definitions) on disk as a B-tree.

The page size has been set at 4KB. This dicatates the minimum size
the file can grow (or shrink) by as necessary. Vacumming is not
necessary since objects are removed from the file directly and empty
pages are pruned at the time.

The algorithm for the B-tree on disk has been built by me from first
principles and stress testing. There are a ton of optimizations that
can (and should) be added soon. I have attached the new benchmark
data and done my best to document (with examples) how the structure and
operations work in the "File Format" documentation.

Perhaps before this refactoring - but only discovered through stress
testing - the default garbage collection mode (autofree) is not yet
stable enough, so you will have to use `boehm`. See "Building from
Source" in the Development documentation for more details.

There is a working memory-based pager included, but there's no way to
utilize it yet. Perhaps this can be used instead of the file-based pager
with a special ":memory:" file name (similar to the way SQLIte3 does it)
but I didn't want to make the decision in this diff, nor throw away the
in-memory implementation.