Allow in-memory database
Opening a database with the file ":memory:" will now use an entirely
in-memory pager. This can also be used with the "bench" by using the
new option `-file :memory` or more simply `make bench-memory`.

Curiously, the updated benchmarks don't show much difference between
on-disk and in-memory, so I guess the OS is doing a very good job with
caching random access reads and writes. Even after making sure all
changes are flushed to disk at the end of each query.

Tests will now use an in-memory database, which on my machine shaves off
about 13% of the execution time.

Added a Makefile to help standardize and simplfy commands for
development and CI.

Fixes #45