Added PRIMARY KEY
A table is allowed to have one optional PRIMARY KEY, although only
one column and integer based types are supported at the moment.

The PRIMARY KEY dictates the underlying B-tree key so rows will
inherently always be brought back in PRIMARY KEY order, even if the
PRIMARY KEY is not used in the query.

The query planner is able to understand very trivial WHERE clauses that
use equality on the PRIMARY KEY, but it's not able to understand
logical operations (such as AND/ORD) or ranges (such as greater than).

Since PRIMARY KEY lookups only need to read the page containing the
index range this leads to an almost 200x speedup over a full table scan.