Added EXPLAIN (#54)
An `EXPLAIN` is not actually a statement, but rather a string prefix
that can be placed before some SQL statements that will cause the query
planner to return the strategy for executing the query without actually
running it. This is useful for debugging query performance.

`EXPLAIN` works for `DELETE`, `UPDATE` and `SELECT`.