Added POSITION() (#66)
POSITION(X IN Y) Returns the start of the left most (first) match of one
string within another. 1 will be the smallest index on a match and 0 is
returned if the substring does not exist. Matching is case-sensitive.

Also added a $TEST environment variable which can be used to run a
single test, like:

    TEST=transaction make sql-test