btree: Support transaction IDs (#68)
This refactors the B-tree and paging implementation from only supporting unique keys across the entire tree to now supporting the created/expired transaction IDs as well as allowing for multiple versions of a key.

This is a breaking change to the file format. I decided to do this as an intermediate step because the diff for transaction was getting really large and complex. Implementing the isolation logic now will be easier and I can feel more confident that the B-tree is well tested.