Add BETWEEN (#72)
Syntax: X [ NOT ] BETWEEN [ ASYMMETRIC | SYMMETRIC ] Y AND Z

A BETWEEN expression returns TRUE if X is inclusively between
Y and Z. Or, put another way (X >= Y AND Z <= Z).

If SYMMETRIC is provided, the orders of Y and Z are irrelevant. Such
that they will be swapped if Y > Z.

If neither ASYMMETRIC or SYMMETRIC is provided then ASYMMETRIC is
assumed.