Add VALUES statement (#73)
The VALUES statement provides one or more rows as a virtual table:

    SELECT * FROM (VALUES 1, 'foo', TRUE);
    SELECT * FROM (VALUES 1, 'foo', TRUE) AS t1 (abc, col2, "f");

But can also be used as s standalone statement:

    VALUES 'cool', 12.3;

This turned out to be far more involved than I originally throught. It
allowed me to overhaul much of the test suite so that we don't need
to use fake tables anymore.