Adding support for uncorrelated subqueries
An uncorrelated subquery is a subquery that does not rely on values from
outside of its scope. This turned out to be a really non-trivial
solution.

Subqueries can only be used in the FROM clause, like:

    SELECT foo AS bar FROM (SELECT x + char_length(y) AS foo FROM t1);

Since this diff is already so massive, I will deal with inline
subqueries (used in expressions) and correlated subqueries as a future
set of work.