F051: Basic date and time
In a nutshell, this adds storage support for date/time types and values.
Although, we do not yet support other features such as constants,
functions, comparisons and casting. They can come as a separate features
in the near future.

More specifically, this adds support for:

1. F051-01: `DATE` data type (including support of `DATE` literal).
2. F051-02: `TIME` data type (including support of `TIME` literal) with
fractional seconds precision of at least 0.
3. F051-03: `TIMESTAMP` data type (including support of `TIMESTAMP`
literal) with fractional seconds precision of at least 0 and 6.

The five new distinct data types and their storage requirements are:

1. `DATE` (8 bytes)
2. `TIME(n) WITH TIME ZONE` (10 bytes)
3. `TIME(n) WITHOUT TIME ZONE` or `TIME(n)` (8 bytes)
4. `TIMESTAMP(n) WITH TIME ZONE` (10 bytes)
5. `TIMESTAMP(n) WITHOUT TIME ZONE` or `TIMESTAMP(n)` (8 bytes)