E021-06: SUBSTRING function
`SUBSTRING` can be constructed in several forms:

  SUBSTRING(
    value
    FROM start_position
    [ FOR string_length ]
    [ USING { CHARACTERS | OCTETS } ]
  )

`start_position` starts at 1 for the first character or byte. If
`start_position` is out of bounds (either before the start or after the
end) the returned value will be empty.

If `string_length` is not provided, all characters or bytes until the
end will be included. Otherwise, only `string_length` will be included.
If `string_length` goes beyond the end of the string it will only be
used until the end.

If `CHARACTERS` is specified the `start_position` and `string_length`
will count in characters (this works with multibyte characters) whereas
`OCTETS` will strictly count in bytes. If `USING` is not provided,
`CHARACTERS` will be used.