F201: CAST function (#122)
The CAST function is used to explicitly change a value to another type.
If the type conversion is not possible, a SQLSTATE 42846, 22001 or 22003
is raised. The last two states are special cases where the conversion is
valid, but the type is not large enough (for strings) or causes an
overflow (for numbers) respectively.

Casting between types might sound fairly trivial, but it's actually is a
major undertaking because up until now vsql has used strict-ish typing.
That is, types that are similar enough the share the same underlying
value (like SMALLINT vs INTEGER vs BIGINT are all stored in an i64
property). This would avoid any overflow checking and worked in a lot of
cases by "coincidence" with extra branches needing to categorize all
these types.

Now, types for operations can be much stricter. Making adding new types
much easier and safer in the future.

- Added SQLSTATE 22001 "string data right truncation". This means a
character value is trying to be converted to a type that is not large
enough to store it. This doesn't just apply to CAST, but also protects
against the same previously unprotected rules for INSERT and UPDATE.

- Added SQLSTATE 42846 "cannot coerce". The general error when a value
cannot be CAST to another type because it is incompatible.

- Fixed bug where CHARCATER(n) types would not zero pad values smaller
than when the length is less than n (as the SQL standard says).

- Many fixes for dates and times that allow conversion between other
date and time types with the correct semantics, rather than the
ultra-strict parsing we previously only allowed for. For example,
inserting a TIMESTAMP WITHOUT TIME ZONE is now permitted and appends the
correct local time zone if inserted into a TIMESTAMP WITH TIME ZONE.

- Lots of documentation cleanup around types and casting rules.