Documentation for V Client Library
There is now a "V Client Library" page to properly document the public
interface when using V to interact with vsql.

To make this easier, a new feature has been created that will extract
"snippets" from comments in V code and put them into `snippets.rst` for
reference in the documentation. This will ensure that we only need to
keep the docs in one place right within the code.