Added `in` and `out` commands (#131)
`vsql in` will execute a SQL file and `vsql out` can be used to dump the
structure and data from a database in SQL format. These are also useful
when converting data to and from other databases.

The commands are designed to be simple and provide a means to make a
lossless copy (or a backup) of a database without any other options:

    vsql out mydb.vsql > mydb.sql
    vsql in mydb.vsql < mydb.sql

This is also especially important if there are changes to the vsql file
format in the future. To be able to quickly and easily recreate the
database.

You can find more CLI options and documentation for these commands here:
https://vsql.readthedocs.io/en/latest/cli.html

To facilitate testing, there is now a 5th test suite called
"CLI Tests" which are written as shell scripts and executed along with
the other suites. This provides a lot of flexibility in testing CLI
options, exit codes, outputs, etc. You can read how CLI Tests work here:
https://vsql.readthedocs.io/en/latest/testing.html#cli