Adding support for SEQUENCEs
A sequence is used as an atomic counter. It is useful for generating
unique and sequential numbers.

The sequence can have an separately optional min and max values, as well
as being able to specify if numbers are allowed to cycle (wrap around)
between these boundaries.

New SQL statements introduced for `CREATE SEQUENCE`, `ALTER SEQUENCE`
and `DROP SEQUENCE` as well as the `NEXT VALUE FOR <name>` expression.

Fixes #139