Adding support for catalogs
A catalog is the SQL standard term for a database. A catalog contains
schemas, but cannot be directly created with SQL. Instead, a catalog in
vsql means a database file which includes a ":memory:" database.

Loading multiple catalogs allows vsql to work seamlessly across multiple
database files, although there is some limitations with transaction
guarantees when working with multiple catalogs. See documentation for
more details.

Catalogs are virtual, meaning that the catalog name is not encoded into
the database file but created when attaching it to the engine. By
default the catalog name is derived from the file by using the first
part of the file name. For example "/foo/bar.baz.vsql" will be "bar".

SQL Standard F651
SQL Standard F762