sequences: Fixing concurrent "NEXT VALUE FOR" (#143)
This fixes a bug where concurrent transactions using "NEXT VALUE FOR"
would make the sequence invisible to the other concurrent transaction.

The properties of a sequence (such as the `INCREMENT BY`, etc) are held
in the same record as the next value. Since the next value of a sequence
needs to be atomic (and separate from the transaction isolation) a
`ROLLBACK` on a transaction that contains an `ALTER SEQUENCE` will not
undo any changes.

Ideally, the properties of a `SEQUENCE` can be stored in a separate
location on disk. However, for now it's a documented limitation.