Various fixes for binary expressions (#165)
The SQL standard states that arithmetic must result in the
"most specific type". This implementation is more correct, but is
incomplete. I'll keep the complete implementation for a later diff after
NUMERIC and DECIMAL has been included.

This also updates predicates to use the SQL-defined version instead of
the existing functions, adds some missing unary operators and fixes a
formatting error for some REAL values.