Overhaul eval into a compiler
This replaces the existing but separate eval(), eval_type() and
reoslve_identifiers() with a much easier and more flexible single
compile() function that encapsulates all aspects of evaluation and
types.

Strictly speaking, this is not a compiler in the traditional sense, but
it does allow us to separate the compile-time logic and the runtime
evaluation with a closure. So apart from simplifying the logic, it
also makes it more efficient by not having to recompile certain parts of
every row.

One more positive side effect from removing separate
resolve_identifier() stage is that we can make the EXPLAIN look a lot
less verbose and natural against the original query.