docs: SQL Standard
Unfortunately, the SQL Standard is not free (and is very expensive).
This includes all previous versions. This is obviously a challenge for
an open-source project. While we cannot include a copy of the standard,
the next best option is to provide a list of locations where it exists.

So far I have been developing against SQL:2016, but we'll slowly roll
this back to SQL:2003.